Add files via upload
In this project, I will be working on the Employees database of a big company. This database consists of 6 tables :-      
departments 
dept_emp 
dept_manager
employees 
salaries 
titles

“https://www.dropbox.com/s/znmjrtlae6vt4zi/employees.sql?dl=0 ”  this is a link for access to the Employees database in SQL script form.

In this project, I work on different questionnaires related to databases, and write queries on My-SQL platform on the basis of these questionnaires and obtain their results. I prefer to show all queries and results with the help of screenshots of My-SQL platform.